MSVC compile fix. Also fix some warnings and trim some empty space.

+++ game/PsyDoom/ImageOps.cpp
@@ -135,95 +135,95 @@ void DebugPrint(
 {
     // Sanity checks and abort if the destination image is zero sized
     ASSERT(pText || (textLen == 0));
-    
+
     if ((dstImg.width == 0) || (dstImg.height == 0))
         return;
-        
+
     ASSERT(dstImg.pPixels);
-    
+
     // Cache these values locally
     PixelT* const pDstPixels = dstImg.pPixels;
     const uint32_t dstW = dstImg.width;
     const uint32_t dstH = dstImg.height;
     const uint32_t dstStride = dstImg.rowStridePx;
-    
+
     // Text printing loop
     int32_t outputCursorX = dstTopLeftX;
     int32_t outputCursorY = dstTopLeftY;
-    
+
     for (uint32_t charIdx = 0; charIdx < textLen; ++charIdx) {
         // Figure out the glyph index and skip the character if it is out of range
         const char rawChar = pText[charIdx];
         const char charUpper = (rawChar >= 'a' && rawChar <= 'z') ? rawChar - 32 : rawChar;
         const int32_t glyphIdx = charUpper - 32;
-        
+
         if ((glyphIdx < 0) || (glyphIdx >= 64))
             continue;
-            
+
         // Handle whitespace characters
         if (charUpper == ' ') {
             outputCursorX += 8;
             continue;
         }
-        
+
         if (charUpper == '\t') {
             outputCursorX += 32;
             continue;
         }
-        
+
         if (charUpper == '\n') {
             outputCursorX = dstTopLeftX;
             outputCursorY += 8;
             continue;
         }
-        
+
         // Which row and column of the font (128x32 @ 4bpp) to get this glyph from?
         // Also compute the top left pixel in the source image.
         const uint32_t glyphRow = (uint32_t) glyphIdx / 16u;
         const uint32_t glyphCol = (uint32_t) glyphIdx % 16u;
-        
+
         constexpr uint32_t SRC_ROW_STRIDE = 64;
         const uint8_t* const pSrcTopLeft = gLIBGPU_DebugFont_Texture + (glyphRow * 8u * SRC_ROW_STRIDE) + (glyphCol * 4u);
-        
+
         // Output the 8x8 pixels of the glyph to the destination image
         for (int32_t glyphY = 0; glyphY < 8; ++glyphY) {
             // Abort if the destination row is out of range
             const int32_t dstY = outputCursorY + glyphY;
-            
+
             if (dstY < 0)
                 continue;
-                
-            if (dstY >= dstH)
+
+            if (dstY >= (int32_t) dstH)
                 break;
-            
+
             // Get the source and destination row
             const uint8_t* const pSrcRow = pSrcTopLeft + glyphY * SRC_ROW_STRIDE;
             PixelT* const pDstRow = &pDstPixels[(uint32_t) dstY * dstStride];
-            
+
             // Output this row of pixels
             for (int32_t glyphX = 0; glyphX < 8; ++glyphX) {
                 // Get whether the source pixel is regarded as opaque/drawn or not and skip over it if not
                 const uint8_t srcByte = pSrcRow[(uint32_t) glyphX / 2u];
                 const uint8_t srcNibble = (glyphX & 1) ? (uint8_t)(srcByte >> 4u) : (uint8_t)(srcByte & 0xF);
                 const bool bOpaquePixel = (srcNibble != 0);
-                
+
                 if (!bOpaquePixel)
                     continue;
-                    
+
                 // Get the destination image x value and skip over this pixel if its out of range
                 const int32_t dstX = outputCursorX + glyphX;
-                
+
                 if (dstX < 0)
                     continue;
-                    
-                if (dstX >= dstW)
+
+                if (dstX >= (int32_t) dstW)
                     break;
-                    
+
                 // Save the destination pixel
                 pDstRow[dstX] = fontColor;
             }
         }
-        
+
         // Move onto the next output location
         outputCursorX += 8;
     }
